apps/backend/pyproject.toml
[tool.poetry]
name = "riskfabric-backend"
version = "0.1.0"
description = "RiskFabric backend service"
authors = ["RiskFabric Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = ">=5.0,<6.0"
djangorestframework = ">=3.15,<4.0"
celery = ">=5.4,<6.0"
redis = ">=5.0,<6.0"
python-dotenv = ">=1.0,<2.0"
PyMySQL = ">=1.1,<2.0"
xlrd = ">=2.0,<3.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0,<9.0"
pytest-django = ">=4.8,<5.0"
ruff = ">=0.9,<1.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
